1.34/background-colors/requirements.txt
streamlit-nightly==1.33.1.dev20240430
spacy>=3.7.2,<3.8.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
spacytextblob